+++ CHANGELOG.md
@@ -14,6 +14,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Removed
 
+- 注文取引結果からeventReservationsを削除(order.acceptedOffersへ移行)
+
 ### Fixed
 
 ### Security